refactor(bridge): extract message actions
